edr_agent/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
pydantic==2.10.2
pydantic_core==2.27.1
python-decouple==3.8
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
